Speech to text conversion
This API can be used by importing 'speech2text'
and invoking speech2text.transcribe().

The API will extract audio from an mp4 file and produce
a raw audio file. This audio file is processed by Google
Cloud to create a textual transcript, which is stored in the
specified output file. If you would like to display the transcript
to the console, set the output file to /dev/stdout